notey-backend/requirements.txt
# Add your Python dependencies here
google-generativeai
python-dotenv
fastapi
pydantic
requests
httpx
